move schema manager
